+++ lmbench2/src/timing.h
@@ -10,7 +10,7 @@ double	Delta(void);
 double	Now(void);
 void	adjust(int usec);
 void	bandwidth(uint64 bytes, uint64 times, int verbose);
-int	bytes(char *s);
+size_t	bytes(char *s);
 void	context(uint64 xfers);
 uint64	delta(void);
 int	get_enough(int);
@@ -41,7 +41,7 @@ void	print_results(int details);
 void	use_int(int result);
 void	use_pointer(void *result);
 uint64	usecs_spent(void);
-void	touch(char *buf, int size);
+void	touch(char *buf, size_t size);
 
 #if defined(hpux) || defined(__hpux) || defined(WIN32)
 int	getpagesize();
